# artifacts/api-server/src/routes/auth.ts
import { Router, Request, Response } from "express";
import { z } from "zod";
import { authMiddleware, AuthenticatedRequest } from "../middlewares/authMiddleware";
import { validateRequest } from "../middlewares/validateRequest";
import { AuthService } from "../services/AuthService";

import rateLimit from "express-rate-limit";

const router = Router();

const authLimiter = rateLimit({
  windowMs: 15 * 60 * 1000,
  max: 10,
  standardHeaders: true,
  legacyHeaders: false,
  message: { error: "Too many auth attempts, please try again later." },
});

const authSchema = z.object({
  body: z.object({
    email: z.string().email(),
    password: z.string().min(10, "Password must be at least 10 characters").max(72, "Password too long"),
    name: z.string().optional()
  }).strict()
});

const resetSchema = z.object({
  body: z.object({
    email: z.string().email()
  }).strict()
});

const confirmResetSchema = z.object({
  body: z.object({
    token: z.string().min(1, "Token is required"),
    new_password: z.string().min(10, "Password must be at least 10 characters")
  }).strict()
});

const logoutSchema = z.object({
  body: z.object({}).strict().optional(),
  headers: z.object({
    authorization: z.string().startsWith("Bearer ")
  }).passthrough()
});

/**
 * @openapi
 * /auth/register:
 *   post:
 *     summary: Register a new user
 *     tags: [Auth]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required: [email, password]
 *             properties:
 *               email:
 *                 type: string
 *               password:
 *                 type: string
 *     responses:
 *       200:
 *         description: User registered successfully
 */
router.post("/register", authLimiter, validateRequest(authSchema), async (req: Request, res: Response) => {
  const result = await AuthService.register(req.body.email, req.body.password);
  res.json(result);
});

/**
 * @openapi
 * /auth/login:
 *   post:
 *     summary: Login an existing user
 *     tags: [Auth]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required: [email, password]
 *             properties:
 *               email:
 *                 type: string
 *               password:
 *                 type: string
 *     responses:
 *       200:
 *         description: Successfully logged in
 */
router.post("/login", authLimiter, validateRequest(authSchema), async (req: Request, res: Response) => {
  const result = await AuthService.login(req.body.email, req.body.password);
  res.json(result);
});

/**
 * @openapi
 * /auth/reset-password:
 *   post:
 *     summary: Request a password reset
 *     tags: [Auth]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required: [email]
 *             properties:
 *               email:
 *                 type: string
 *     responses:
 *       200:
 *         description: Reset instructions sent
 */
router.post("/reset-password", authLimiter, validateRequest(resetSchema), async (req: Request, res: Response) => {
  const result = await AuthService.resetPassword(req.body.email);
  res.json(result);
});

/**
 * @openapi
 * /auth/logout:
 *   post:
 *     summary: Logout user
 *     tags: [Auth]
 *     security:
 *       - bearerAuth: []
 *     responses:
 *       200:
 *         description: Logged out successfully
 */
router.post("/logout", authMiddleware, validateRequest(logoutSchema), async (req: AuthenticatedRequest, res: Response) => {
  const result = await AuthService.logout(req.headers.authorization, req.user?.sub);
  res.json(result);
});

/**
 * @openapi
 * /auth/confirm-reset:
 *   post:
 *     summary: Confirm password reset
 *     tags: [Auth]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required: [token, new_password]
 *             properties:
 *               token:
 *                 type: string
 *               new_password:
 *                 type: string
 *     responses:
 *       200:
 *         description: Password reset successful
 */
router.post("/confirm-reset", authLimiter, validateRequest(confirmResetSchema), async (req: Request, res: Response) => {
  const result = await AuthService.confirmReset(req.body.token, req.body.new_password);
  res.json(result);
});

/**
 * @openapi
 * /auth/refresh:
 *   post:
 *     summary: Refresh auth token
 *     tags: [Auth]
 *     security:
 *       - bearerAuth: []
 *     responses:
 *       200:
 *         description: Token refreshed
 */
router.post("/refresh", authMiddleware, async (req: AuthenticatedRequest, res: Response) => {
  // authMiddleware already validates the token. We just issue a new one.
  const user = req.user;
  if (!user) {
    res.status(401).json({ error: "Unauthorized" });
    return;
  }
  
  const secret = process.env.SUPABASE_JWT_SECRET?.split(',')[0];
  if (!secret) {
    res.status(500).json({ error: "Server configuration error" });
    return;
  }

  const jwt = require("jsonwebtoken");
  const newToken = jwt.sign({
    sub: user.sub,
    email: user.email,
    role: user.role,
    aud: user.aud
  }, secret, { expiresIn: '1h' });

  res.json({
    session: { access_token: newToken, expires_in: 3600 }
  });
});

export default router;